Implement filter and count feature
The program now checks each line from the input file and compares it
to the previous line. It then only prints out the lines that repeat.

It also keeps track of how many times each line repeats and the
format_count function prints it if the count flag is specified.

Changed the method of reading from read_line() to lines() to mimic
uniq's behavior regarding newline bytes. read_line() keeps the newline
while lines() does not. uniq does not differentiate lines with different
newlines so lines() is used.