Refactor output to remove vector
Previously, we moved the entire output to a string vector, then printed
each string into either a file or STDOUT depending on the output argument.
Now, we just write each line directly to an out_file variable, which is
either a new File or STDOUT depending on the argument. This way we
don't have to create/use the vector.

The empty file check now happens before writing the last line. Before,
it didn't matter since we were writing the last line to a vector and
it would check before printing out the vector. Now, since it writes
directly, the check must take place before the last line write.